feat(oxfmt): Rename embedded.ts with preparing formatFile() function
