+++ apps/oxfmt/src-js/index.ts
@@ -1,2 +1,2 @@
 export * from "./bindings.js";
-export { formatEmbeddedCode } from "./embedded.js";
+export { formatEmbeddedCode, formatFile } from "./prettier-proxy.js";
